Add a testcase and boilerplate code to get a user identity from a handle.
We are going to provide some kind of user identity using the Gravatar
service, as this seems to be the easiest way to get started (for now).

Create the test case and the module boilerplate code. Do not forget to
add the @doc attributes to the code.

Run the tests by typing the following at the terminal:
> mix do compile, test

Please note that this test will fail for now, as the functionality is not
yet implemented.